Fix prubin installation
The `/usr/lib/linuxcnc/prubin` files and the symlinks in
`/usr/lib/linuxcnc/$(flavor)` are installed only through some
impressive calisthenics in the Debian packaging (and a surprise assist
from autoconf).  However, they are not installed as expected with
`make install`.

This patch reverts to tradition, installing the files and symlinks
with `make install`, and adding those results to the Debian packaging.
The old scheme is cleaned up from packaging and autoconf.