rtapi_support.c:  Temp. silence `rtapi_print()`
This reverts part of commit 77a23d3 from PR #1452.

The `rtapi_print()` function was set to never print, and that PR
changed it to always print, as the comments in `rtapi.h` indicate it
should.  However, the result was that it spews messages out the error
channel, which show up in Axis.  That's bad.

This reverts back to the "never print" behavior until a better
solution comes along.