Revamp MKUUID generation
The `configure.ac` macros generated a default MKUUID and created the
`../etc/machinekit/machinekit.ini` file.  MKUUID needs to be generated
at install-time for packages or build-time for running out of tree.

As a result, running `make clean` removed `machinekit.ini`, but `make`
didn't replace it, breaking runs.

It also made managing the MKUUID very complicated, as evidenced by
complicated scripts and a history showing them to be delicate.

This patch fixes that by generating `machinekit.ini` (with new MKUUID)
from the `Makefile` for out-of-tree builds, and simplifying packaging
by installing the `makefile.ini` template and generating the MKUUID in
the `.postinst` script.