MAINTENANCE: Change test for <sys/io.h> header file
The HAS_SYS_IO and SYS_IO_DEFINED switches were up to now dependent on the HOST architecture type (allow on all but 'arm64'). This actually caused problems on Ubuntu Focal 20.04 LTS, where libc6-dev package on 'armhf' platform does not include the <sys/io.h> file.

This change use AC_CHECK_HEADER macro in Autoconf script to determine if the system has or has not the header.

Patch includes minimal change to the build-flow (i.e. does not care if this is the best macro for the task, only to cause the least amount of ripple).