tests/threads.0:  Add instrumentation to prove preemption
From https://github.com/machinekit/machinekit-hal/issues/173#issuecomment-684166252:

The original purpose of this test in the LCNC project was to
demonstrate that HAL threads (which run on a single CPU; LCNC-HAL
doesn't have multi-core) have rate-monotonic scheduling, i.e. the
`slow` thread will never preempt the `fast` thread.  I've done some
work to show that these test failures are due either to lack of
rate-monotonic scheduling, when threads fail to set `SCHED_FIFO`, or
else because they are running in parallel on multiple CPUs, and
scheduling is not to blame.

I added two types instrumentation along the lines of
https://github.com/machinekit/machinekit-hal/issues/173#issue-397211641
and ran it on a single core machine.

Mutex:  The original `fast` thread function sets a mutex, the
`sampler` function executes, and then a (new) second function on the
`fast` thread clears the mutex.  The original `reset` function on the
`slow` thread checks whether the mutex is set; the intention is to
further prove that the `count` pin is reset to zero because the `slow`
thread is preempting the `fast` thread.

Timestamps:  Each of the three functions now writes its start and stop
time on new pins.  Each of these three functions is followed by its
own dedicated `sampler` instance that records the function's start and
stop time, the mutex state, and of course the `count`.  The
`checkresult` massages all the timestamps from these three sources
together, in order, where it again becomes plain from the respective
start and stop timestamps when the `slow` thread has preempted the
`fast` thread .

The reason I went through all this apparently redundant trouble was to
convince myself that the failures weren't due to something else like
memory barrier leakage, and now I am quite convinced.