scripts/realtime:  Better cleanup after failed shutdown
The `hm2_idrom` tests fail frequently.  I caught that red-handed with
a `ps -ef` added to the test script and found three stale `rtapi:0`
and a stale `msgd:0` process lingering around, and a `halcmd ping`
core dump that appeared to have tried to access an old shm segment
that was removed.

This patch gives `rtapi_msgd` processes a little extra time to exit
gracefully, tries to clean up any stuck `rtapi_app` processes
afterwards, and exits non-zero if it detects it has failed.