posix & rt-preempt threads:  Fix RT deadline exceptions
The logic for handling RTP_DEADLINE_MISSED was reversed:  ignored for
rt-preempt, reported for posix!  Because of this, posix reported
missed deadlines:

  1:rtapi_app:11466:user 1: unspecified exception detail=0xd5f754 ts=0xb5d81154

The exception is 'unspecified' because there was no exception handler
for posix threads.  (This is now added, even though it won't be used.)

Also RT flag logic was broken, but not in a way that would affect
operation.